pyproject.toml
# Helps scripts below work
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# Required when using src layout
[tool.hatchling.build.targets.wheel]
packages = ["src/sync"]

[project]
name = "sync"
version = "0.1.0"
description = "Adds my YouTube subscriptions to my Feedbin subscriptions and my liked videos to Notion."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-api-python-client>=2.187.0",
    "google-auth-oauthlib>=1.2.3",
    "httpx>=0.28.1",
    "logfire>=4.18.0",
    "opentelemetry-instrumentation-httpx>=0.60b1",
    "opentelemetry-instrumentation-requests>=0.60b1",
    "opentelemetry-instrumentation-sqlite3>=0.60b1",
    "opentelemetry-instrumentation-urllib>=0.60b1",
    "pydantic>=2.12.5",
    "result>=0.17.0",
    "rich>=14.2.0",
    "structlog>=25.5.0",
]

[dependency-groups]
dev = [
    "google-api-python-client-stubs>=1.31.0",
    "inline-snapshot[dirty-equals]>=0.31.1",
    "pytest>=9.0.2",
    "pytest-httpx>=0.36.0",
    "pytest-watcher>=0.6.3",
    "ruff>=0.14.10",
    "ty>=0.0.8",
]
